+++ Marlin/src/HAL/STM32F1/HAL.h
@@ -183,31 +183,6 @@ typedef int8_t pin_t;
 // ADC
 // ------------------------
 
-enum ADCIndex : uint8_t {
-  OPTITEM(HAS_TEMP_ADC_0, TEMP_0)
-  OPTITEM(HAS_TEMP_ADC_1, TEMP_1)
-  OPTITEM(HAS_TEMP_ADC_2, TEMP_2)
-  OPTITEM(HAS_TEMP_ADC_3, TEMP_3)
-  OPTITEM(HAS_TEMP_ADC_4, TEMP_4)
-  OPTITEM(HAS_TEMP_ADC_5, TEMP_5)
-  OPTITEM(HAS_TEMP_ADC_6, TEMP_6)
-  OPTITEM(HAS_TEMP_ADC_7, TEMP_7)
-  OPTITEM(HAS_HEATED_BED, TEMP_BED)
-  OPTITEM(HAS_TEMP_CHAMBER, TEMP_CHAMBER)
-  OPTITEM(HAS_TEMP_ADC_PROBE, TEMP_PROBE)
-  OPTITEM(HAS_TEMP_COOLER, TEMP_COOLER)
-  OPTITEM(HAS_TEMP_BOARD, TEMP_BOARD)
-  OPTITEM(HAS_TEMP_SOC, TEMP_SOC)
-  OPTITEM(FILAMENT_WIDTH_SENSOR, FILWIDTH)
-  OPTITEM(HAS_ADC_BUTTONS, ADC_KEY)
-  OPTITEM(HAS_JOY_ADC_X, JOY_X)
-  OPTITEM(HAS_JOY_ADC_Y, JOY_Y)
-  OPTITEM(HAS_JOY_ADC_Z, JOY_Z)
-  OPTITEM(POWER_MONITOR_CURRENT, POWERMON_CURRENT)
-  OPTITEM(POWER_MONITOR_VOLTAGE, POWERMON_VOLTS)
-  ADC_COUNT
-};
-
 #ifdef ADC_RESOLUTION
   #define HAL_ADC_RESOLUTION ADC_RESOLUTION
 #else
