+++ Marlin/src/HAL/STM32F1/HAL_N32.cpp
@@ -31,6 +31,7 @@
 #include "HAL_N32.h"
 #include "HAL.h"
 
+#include "adc.h"
 static uint16_t adc_results[ADC_COUNT];
 
 void ADC_Init(ADC_Module* NS_ADCx, ADC_InitType* ADC_InitStruct) {
